Fix titlepage not working
For some absolutely bizarre reason, loading hyperdoc will stop the
titlepage from working when the option is provided to the package.  I
have no idea why, and it was very hard to find, but that's the issue.
Technically the issue was introduced in db1c2c8 but I previously had
hyperdoc so I'm not really sure what happened there.  Perhaps it was
the load order as hyperdoc was previously loaded before the statement
telling the class to \maketitle at the start of the document; but I
don't know (and I kind of don't care now that it is working).
Hopefully the last little bug with regard to the big refactor, because
these are a little time consuming!